+++ scripts/combine_snvs.sh
@@ -28,6 +28,7 @@ if [[ $LG3_DEBUG ]]; then
 fi
 
 
+assert_python ""
 unset PYTHONPATH  ## ADHOC: In case it is set by user
 
 ### Input