+++ scripts/libID_to_patientID.sh
@@ -28,6 +28,8 @@ if [[ $LG3_DEBUG ]]; then
 fi
 
 PROG=$(basename "$0")
+
+assert_python ""
 unset PYTHONPATH  ## ADHOC: In case it is set by user
 
 CONV=$1