+++ scripts/pindel_filter.sh
@@ -35,6 +35,8 @@ fi
 ##
 #
 
+assert_python ""
+
 BEDTOOLS=/opt/BEDTools/BEDTools-2.16.2/bin/bedtools
 PYTHON_PINDEL_FILTER=${LG3_HOME}/scripts/pindel_filter.py
 assert_file_executable "${BEDTOOLS}"